pyproject.toml
[project]
name = "rental-aggregator"
version = "0.2.0"
description = "House Hunting Agent — Multi-source rental listing aggregator with dedup, ranking, and broker outreach"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "googlemaps>=4.10.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "apscheduler>=3.10.0",
]

[project.scripts]
rental = "main:main"
